Political-Chrome-Extension/requirements.txt
azure-functions
azure-ai-textanalytics
newspaper3k
twilio